Improve main menu spline issue
This was due to the moveDelta rotation being to fast whilst cancelling
the spline. Instead we find the direction but using the path behind us.
As long as the precision isn't too high (ideally 0.1-0.02f max), then
this will look smooth, and cancelle spline animations will complete
without issue.

There's still a slight difference in spline output when doing the
following:
- Click on Leaderboard submenu (small spline) - character in place
- Click back
- Click on How to play
- Click back
- Click on Leaderboard submenu - character slightly offset