Fix menu button colours
Translucent buttons are nice but don't work on everything, flat can look
substancially better, so the default has changed to flat and using blur
for submenu background / hint text only.

I've slightly changed the colors and will stop here. It's probably the
best it will get.

This fixes the grey icons in a death screen and makes interface buttons
easier to read.