Add TimeAddedWidget
This fixes a minus issue with UIManager.formatTime and adds a ToFixed()
helper extension for double/float/decimal similar to JS (sometimes to use
than String.Format()).

This can use green/red styling for +/- respectively but this has been
temporarily commented out in favour of being able to style the row
yourself.