Improvements to TimeAddedWidget
Time notifications will be considered expired when their opacity is
approximately 0.1f to help keep the notification stack cleared and will
hardly be mostly invisible by this point.

Additionally we now support modifying the notification if the TimeSpan
to amend has the same sign. This setting is off by default.

When the capacity is too high it's best to simply pop without needing to
check expiry. During removal we will take from the top first. I believe
check in BuildHash is still safe to modify since the render must be
proven dirty first.